+++ nhais-adaptor/edifact/incoming/parser/message.py
